+++ src/app/Fake.Core.Trace/TraceListener.fs
@@ -8,27 +8,101 @@ open System
 type KnownTags =
     | Task of name:string
     | Target of name:string
+    | Compilation of compiler:string
+    | TestSuite of suiteName:string
+    | Test of testName:string
     | Other of typeDef:string * name:string
     member x.Name =
         match x with
         | Task n
         | Target n
+        | Compilation n
+        | TestSuite n
+        | Test n
         | Other (_, n) -> n
     member x.Type =
         match x with
         | Task _ -> "task"
         | Target _ -> "target"
+        | Compilation _ -> "compilation"
+        | TestSuite _ -> "testsuite"
+        | Test _ -> "test"
         | Other (t, _) -> t
+
+type DotNetCoverageTool =
+    | DotCover
+    | PartCover
+    | NCover
+    | NCover3
+    override x.ToString() =
+        match x with
+        | DotCover -> "dotcover"
+        | PartCover -> "partcover"
+        | NCover -> "ncover"
+        | NCover3 -> "ncover3"
+
+type ImportData =
+    | BuildArtifact
+    | DotNetCoverage of DotNetCoverageTool
+    | DotNetDupFinder
+    | PmdCpd
+    | Pmd
+    | FxCop
+    | ReSharperInspectCode
+    | Jslint
+    | FindBugs
+    | Checkstyle
+    | Gtest
+    | Mstest
+    | Surefire
+    | Junit
+    | Nunit
+    member x.Name =
+        match x with
+        | DotNetCoverage _ -> "dotNetCoverage"
+        | DotNetDupFinder -> "DotNetDupFinder"
+        | PmdCpd -> "pmdCpd"
+        | Pmd -> "pmd"
+        | ReSharperInspectCode -> "ReSharperInspectCode"
+        | Jslint -> "jslint"
+        | FindBugs -> "findBugs"
+        | Checkstyle -> "checkstyle"
+        | Gtest -> "gtest"
+        | Mstest -> "mstest"
+        | Surefire -> "surefire"
+        | Junit -> "junit"
+        | FxCop -> "FxCop"
+        | Nunit -> "nunit"
+    override x.ToString() =
+        match x with
+        | DotNetCoverage tool -> sprintf "dotNetCoverage (%O)" tool
+        | _ -> x.Name
+
+type TestStatus =
+    | Ignored
+    | Failed of message:string * details:string * expectedActual:(string * string) option
+
+module TestStatus =
+    let inline mapMessage f (t:TestStatus) =
+        match t with
+        | Failed (message, details, Some (expected, actual)) ->
+            Failed (f message, f details, Some (f expected, f actual))
+        | Failed (message, details, None) ->
+            Failed (f message, f details, None)
+        | _ -> t        
+
 
 /// Defines Tracing information for TraceListeners
 type TraceData = 
-    | StartMessage
-    | ImportantMessage of string
-    | ErrorMessage of string
-    | LogMessage of string * bool
-    | TraceMessage of string * bool
-    | FinishedMessage
+    | ImportData of typ:ImportData * path:string
+    | BuildNumber of text:string
+    | ImportantMessage of text:string
+    | ErrorMessage of text:string
+    | LogMessage of text:string * newLine:bool
+    | TraceMessage of text:string * newLine:bool
     | OpenTag of KnownTags * description:string
+    | TestStatus of testName:string * status:TestStatus
+    | TestOutput of testName:string * out:string * isStdErr:bool
     | CloseTag of KnownTags
     member x.NewLine =
         match x with
@@ -36,8 +110,10 @@ type TraceData =
         | ErrorMessage _ -> Some true
         | LogMessage (_, newLine)
         | TraceMessage (_, newLine) -> Some newLine
-        | StartMessage
-        | FinishedMessage
+        | BuildNumber _
+        | TestStatus (_,_)
+        | TestOutput (_,_,_)
+        | ImportData (_,_)
         | OpenTag _
         | CloseTag _ -> None
     member x.Message =
@@ -46,8 +122,10 @@ type TraceData =
         | ErrorMessage text
         | LogMessage (text, _)
         | TraceMessage (text, _) -> Some text
-        | StartMessage
-        | FinishedMessage
+        | BuildNumber _
+        | TestStatus (_,_)
+        | TestOutput (_,_,_)
+        | ImportData (_,_)
         | OpenTag _
         | CloseTag _ -> None
 
@@ -58,10 +136,9 @@ module TraceData =
         | ErrorMessage text -> ErrorMessage (f text)
         | LogMessage (text, d) -> LogMessage (f text, d)
         | TraceMessage (text, d) -> TraceMessage (f text, d)
-        | StartMessage -> StartMessage
-        | FinishedMessage -> FinishedMessage
-        | OpenTag (t, d) -> OpenTag (t, d)
-        | CloseTag d -> CloseTag d
+        | TestStatus (testName,status) -> TestStatus(testName, TestStatus.mapMessage f status)
+        | TestOutput (testName,out,isStdErr) -> TestOutput (testName,f out,isStdErr)
+        | _ -> t
 
     let internal repl (oldStr:string) (repl:string) (s:string) =
         s.Replace(oldStr, repl)
@@ -72,12 +149,9 @@ module TraceData =
 type ITraceListener = 
     abstract Write : TraceData -> unit
 
-/// Implements a TraceListener for System.Console.
-/// ## Parameters
-///  - `importantMessagesToStdErr` - Defines whether to trace important messages to StdErr.
-///  - `colorMap` - A function which maps TracePriorities to ConsoleColors.
-type ConsoleTraceListener(importantMessagesToStdErr, colorMap) = 
-    let writeText toStdErr color newLine text = 
+module ConsoleWriter =
+    
+    let write toStdErr color newLine text = 
         let curColor = Console.ForegroundColor
         try
           if curColor <> color then Console.ForegroundColor <- color
@@ -91,6 +165,21 @@ type ConsoleTraceListener(importantMessagesToStdErr, colorMap) =
         finally
           if curColor <> color then Console.ForegroundColor <- curColor
 
+    /// A default color map which maps TracePriorities to ConsoleColors
+    let colorMap traceData = 
+        match traceData with
+        | ImportantMessage _ -> ConsoleColor.Yellow
+        | ErrorMessage _ -> ConsoleColor.Red
+        | LogMessage _ -> ConsoleColor.Gray
+        | TraceMessage _ -> ConsoleColor.Green
+        | FinishedMessage -> ConsoleColor.White
+        | _ -> ConsoleColor.Gray
+
+/// Implements a TraceListener for System.Console.
+/// ## Parameters
+///  - `importantMessagesToStdErr` - Defines whether to trace important messages to StdErr.
+///  - `colorMap` - A function which maps TracePriorities to ConsoleColors.
+type ConsoleTraceListener(importantMessagesToStdErr, colorMap) =
     interface ITraceListener with
         /// Writes the given message to the Console.
         member this.Write msg = 
@@ -100,17 +189,15 @@ type ConsoleTraceListener(importantMessagesToStdErr, colorMap) =
             | OpenTag _ -> ()
             | CloseTag _ -> ()
             | ImportantMessage text | ErrorMessage text ->
-                writeText importantMessagesToStdErr color true text
+                ConsoleWriter.write importantMessagesToStdErr color true text
             | LogMessage(text, newLine) | TraceMessage(text, newLine) ->
-                writeText false color newLine text
+                ConsoleWriter.write false color newLine text
             | FinishedMessage -> ()
 
 type TraceSecret =
     { Value : string; Replacement : string }
 
 module TraceSecrets =
-
-
     let private traceSecretsVar = "Fake.Core.Trace.TraceSecrets"
     let private getTraceSecrets, _, (setTraceSecrets:TraceSecret list -> unit) = 
         Fake.Core.Context.fakeVar traceSecretsVar
@@ -128,21 +215,12 @@ module TraceSecrets =
         |> Seq.fold (fun state secret -> TraceData.repl secret.Value secret.Replacement state) s
 
 module CoreTracing =
-    /// A default color map which maps TracePriorities to ConsoleColors
-    let colorMap traceData = 
-        match traceData with
-        | ImportantMessage _ -> ConsoleColor.Yellow
-        | ErrorMessage _ -> ConsoleColor.Red
-        | LogMessage _ -> ConsoleColor.Gray
-        | TraceMessage _ -> ConsoleColor.Green
-        | FinishedMessage -> ConsoleColor.White
-        | _ -> ConsoleColor.Gray
     // If we write the stderr on those build servers the build will fail.
     let importantMessagesToStdErr = buildServer <> CCNet && buildServer <> AppVeyor && buildServer <> TeamCity && buildServer <> TeamFoundation
 
     /// The default TraceListener for Console.
     let defaultConsoleTraceListener  =
-      ConsoleTraceListener(importantMessagesToStdErr, colorMap) :> ITraceListener
+      ConsoleTraceListener(importantMessagesToStdErr, ConsoleWriter.colorMap) :> ITraceListener
 
 
     /// A List with all registered listeners
@@ -151,12 +229,16 @@ module CoreTracing =
     let private getTraceListeners, _, (setTraceListenersPrivate:ITraceListener list -> unit) = 
         Fake.Core.Context.fakeVar traceListenersVar
 
+    let areListenersSet () =
+        match getTraceListeners() with
+        | None -> false
+        | Some _ -> true
+
+
     // register listeners
     let getListeners () =
         match getTraceListeners() with
-        | None ->
-            setTraceListenersPrivate [defaultConsoleTraceListener]
-            [defaultConsoleTraceListener]
+        | None -> [defaultConsoleTraceListener]
         | Some t -> t
 
     let setTraceListeners l = setTraceListenersPrivate l